ocr_service/requirements.txt
fastapi
uvicorn
pillow
pytesseract
